Split data structs out to separate module and make fields public.
